Update bufpage with a bit of data layout, the rest can only be done after type and tuple are implemented

+++ src/storage/mod.rs
@@ -1,4 +1,4 @@
-pub const PAGE_SIZE:u64 = 4096;
+pub const PAGE_SIZE:u32 = 4096;
 
 pub mod bufkey;
 pub mod bufmgr;
